server/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
Authlib==1.6.5
beautifulsoup4==4.14.2
blinker==1.9.0
blis==1.3.0
cachetools==6.2.0
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.22.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
cryptography==45.0.7
curl_cffi==0.13.0
cycler==0.12.1
cyclopts==3.24.0
cymem==2.0.11
databricks-sdk==0.67.0
datasets==4.1.1
dateparser==1.2.2
decorator==5.2.1
dill==0.4.0
distlib==0.3.9
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.2
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.118.0
fastmcp==2.12.4
filelock==3.16.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.60.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.7.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyterlab_widgets==3.0.15
keras==3.11.3
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lazy-object-proxy==1.12.0
libclang==18.1.1
llvmlite==0.45.1
Mako==1.3.10
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mcp==1.16.0
mdurl==0.1.2
ml_dtypes==0.5.3
mlflow==3.4.0
mlflow-skinny==3.4.0
mlflow-tracing==3.4.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
multitasking==0.0.12
murmurhash==1.0.13
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.6.0
networkx==3.5
nltk==3.9.2
numba==0.62.1
numpy==2.3.2
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
parse==1.20.2
parso==0.8.5
pathable==0.4.4
pbr==6.1.0
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.6
plotly==6.3.1
preshed==3.0.10
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyngrok==7.4.0
pyparsing==3.2.5
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
schedule==1.2.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
sqlparse==0.5.3
srsly==2.5.1
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
stevedore==5.4.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0+cpu
torchvision==0.23.0+cpu
toutatis==1.31
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.2
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
vectorbt==0.28.1
virtualenv==20.28.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
virtualenvwrapper-win==1.2.7
waitress==3.0.2
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.1
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.3
xgboost==3.0.5
xxhash==3.6.0
yarl==1.20.1
yfinance==0.2.66
zipp==3.23.0
